moved preload.js
